refactor: move useMenuItem to its own folder.
