+++ docs/data/base/components/menu/menu-zh.md
@@ -105,7 +105,7 @@ Note that `componentsProps` slot names are written in lowercase (`root`) while `
 
 ```jsx
 import { useMenu } from '@mui/base/MenuUnstyled';
-import { useMenuItem } from '@mui/base/MenuItemUnstyled';
+import useMenuItem from '@mui/base/useMenuItem';
 ```
 
 The `useMenu` and `useMenuItem` hooks let you apply the functionality of the menu components to fully custom components. They return props to be placed on the custom components, along with fields representing the components' internal states.
