+++ docs/data/base/components/menu/menu.md
@@ -115,7 +115,7 @@ Note that `componentsProps` slot names are written in lowercase (`root`) while `
 
 ```jsx
 import { useMenu } from '@mui/base/MenuUnstyled';
-import { useMenuItem } from '@mui/base/MenuItemUnstyled';
+import useMenuItem from '@mui/base/useMenuItem';
 ```
 
 The `useMenu` and `useMenuItem` hooks let you apply the functionality of the menu components to fully custom components.
