+++ packages/mui-base/src/MenuItemUnstyled/index.tsx
@@ -4,6 +4,3 @@ export * from './MenuItemUnstyled.types';
 
 export { default as menuItemUnstyledClasses } from './menuItemUnstyledClasses';
 export * from './menuItemUnstyledClasses';
-
-export { default as useMenuItem } from './useMenuItem';
-export * from './useMenuItem.types';
