+++ packages/mui-base/src/useMenuItem/index.ts
@@ -1,0 +1,5 @@
+export { default as useMenuItem } from './useMenuItem';
+export { default } from './useMenuItem';
+export * from './useMenuItem';
+
+export * from './useMenuItem.types';
